[Fix] Updated User Table Page to MVC
- Made userTable components (controller & view)
- Cleaned dashboard view column
